+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Imp.hs
@@ -25,6 +25,7 @@ import Cardano.Ledger.Shelley.Rules (
  )
 import qualified Test.Cardano.Ledger.Alonzo.Imp as AlonzoImp
 import Test.Cardano.Ledger.Alonzo.ImpTest (AlonzoEraImp, LedgerSpec)
+import qualified Test.Cardano.Ledger.Babbage.Imp.UtxoSpec as Utxo
 import qualified Test.Cardano.Ledger.Babbage.Imp.UtxowSpec as Utxow
 import Test.Cardano.Ledger.Imp.Common
 
@@ -47,3 +48,4 @@ spec = do
   AlonzoImp.spec @era
   describe "BabbageImpSpec" . withImpInit @(LedgerSpec era) $ do
     Utxow.spec
+    Utxo.spec